requirements.txt
# Google API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Phone number parsing
phonenumbers>=8.13.0

# Fuzzy string matching
rapidfuzz>=3.5.0

# Unicode/diacritics
unidecode>=1.3.7

# Name parsing
nameparser>=1.1.2

# Email validation
email-validator>=2.1.0

# Progress bars
tqdm>=4.66.0

# Rich terminal output
rich>=13.7.0
